Reorganize test files
