+++ tests/auth/test_permission.py
@@ -101,9 +101,14 @@ class PermissionTestCase(ApiDBTestCase):
         self.generate_fixture_asset_type()
         self.generate_fixture_asset()
         self.generate_assigned_task()
-        tasks_service.assign_task(self.task.id, self.user_cg_artist.id)
+        task_id = str(self.task.id)
         self.log_in_cg_artist()
-        self.get("data/projects", 403)
+        user_id = str(self.user_cg_artist.id)
+        projects = self.get("data/projects")
+        self.assertEquals(len(projects), 0)
+        tasks_service.assign_task(task_id, user_id)
+        projects = self.get("data/projects")
+        self.assertEquals(len(projects), 1)
         projects = self.get("data/projects/all")
         self.assertEquals(len(projects), 1)
         projects = self.get("data/projects/open")
