+++ tests/shots/test_breakdown.py
@@ -1,5 +1,7 @@
 from tests.base import ApiDBTestCase
 
+from zou.app.models.entity import Entity
+
 
 class BreakdownTestCase(ApiDBTestCase):
 
@@ -64,3 +66,21 @@ class BreakdownTestCase(ApiDBTestCase):
         self.assertEquals(cast_in[0]["shot_name"], self.shot.name)
         self.assertEquals(cast_in[0]["sequence_name"], self.sequence.name)
         self.assertEquals(cast_in[0]["episode_name"], self.episode.name)
+
+    def test_get_assets_for_shots(self):
+        self.entities = self.generate_data(
+            Entity, 3,
+            entities_out=[],
+            entities_in=[],
+            project_id=self.project.id,
+            entity_type_id=self.asset_type.id
+        )
+        self.shot.entities_out = self.entities
+        self.shot.save()
+
+        assets = self.get("data/shots/%s/assets" % self.shot.id)
+        self.assertEquals(len(assets), 3)
+        self.assertDictEqual(
+            assets[0],
+            self.entities[0].serialize(obj_type="Asset")
+        )
