+++ tests/source/shotgun/test_shotgun_import_project.py
@@ -11,7 +11,7 @@ class ImportShotgunProjectTestCase(ShotgunTestCase):
         super(ImportShotgunProjectTestCase, self).setUp()
 
     def test_import_projects(self):
-        self.projects = self.load_fixture('projects')
+        self.projects = self.load_fixture("projects")
         self.assertEqual(len(self.projects), 2)
 
         self.projects = self.get("data/projects")
@@ -28,8 +28,8 @@ class ImportShotgunProjectTestCase(ShotgunTestCase):
         self.assertDictEqual(self.projects[0]["file_tree"], tree)
 
     def test_import_projects_twice(self):
-        self.projects = self.load_fixture('projects')
-        self.projects = self.load_fixture('projects')
+        self.projects = self.load_fixture("projects")
+        self.projects = self.load_fixture("projects")
         self.assertEqual(len(self.projects), 2)
 
         self.projects = self.get("data/projects")
